renaming / code review
