+++ src/services/ContributorAuditService/providers/GitHub/index.ts
@@ -1,0 +1,2 @@
+export * from "./GitHubApiClient";
+export * from "./GitHubContributorAuditProvider";
