--- src/services/ContributorAuditService/scm/GitHub/api/index.ts
@@ -1,1 +1,0 @@
-export * from "./GitHubApiClient";
