--- src/services/ContributorAuditService/scm/GitHub/index.ts
@@ -1,2 +1,0 @@
-export * from "./GitHubAudit";
-export * from "./utilities";
